requirement.txt
appdirs==1.4.4
asn1crypto==1.5.1
attrs==23.1.0
beautifulsoup4==4.12.2
cached-property==1.5.2
certifi==2023.7.22
cffi==1.15.1
colorama==0.4.6
defusedxml==0.7.1
Django==2.2
django-cors-headers==2.5.3
djangorestframework==3.9.2
importlib-metadata==6.7.0
isodate==0.6.1
Pillow==9.5.0
pycparser==2.21
pyparsing==3.1.1
pytz==2023.3.post1
pywin32==306
requests-toolbelt==1.0.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
zipp==3.15.0
